GUI on right thumb
On the kyria I had a whole I3WM layer which mostly mapped gui(kc)
combinations. I'm trying to avoid such layer here on the corne. Having a
gui kc instead of a dedicated layer could be a bit less optimized, but
will force me to use a more standard way of handling i3wm and maybe to
rely on software configuration rather than keyboard config for
optimization. With the current solution (dedicated layer) it was quite
difficult to use i3wm with a standard keyboard.