Define number layer
I kept numbers on the left side as I have on the kyria. I've lost the
numpad like configuration on the right side, as I've never really used
it. Instead I've repeated the right side of the symbol layers given most
of them are numeric related and it could make a lot of sense to have
them here. We'll see.
Also I kept in the num layer the bluetooth pairing keys, since I had the
space.