Compact layers to remove nav
This is backported from an experiment I am doing with the kyria.
The idea is to try to avoid a third (semi empty) layer for the
navigation. I was not using the right half of the number layer anyway.

This also fixes some minor differencies between kyria and cornish zen
layout.